Setup files to write Ada frontend
Infer IR is using an imperative structure for the control flow graph.
It is not very conveninant to directly translate to this CFG.

Instead, use an intermediate representation that describes the CFG.
Thus, statements and expressions are first translate to this IR.

Implement some expressions and statements using the intermediate
representation.

TN: S410-042