Implement ExtendedReturnStmt
To be able to implement this, we modify the context so that it carries
a substution map.

Implement the function "pvar" that for a given ada node, return either what
is in the substitution map for the defining name referenced by the node, or
creates a program variable using the defining name and the computation
of it's unique name.

TN: S410-042