Refactor assignments to use an external function
This will allow us to better control the checks that we want to add for
an assignment.

For the checks we need the lal type expression of the assigned value.
Thus, pass this to the trans_assignment function.

In order to call trans_assignement for a ReturnStmt, we need to add the
returned type expression in the context.

TN: S410-042