Implement parameter passing strategies
Add two strategies, Copy and Reference.

Add a map from DefiningName to the strategy in the context.

When evaluating an lvalue, look in this map to see if the expression
denotes a parameter and load it as it would be a pointer.

trans_call should be called with a pair with the mode and the
expression. Modify calls to trans_call to take that into account.

TN: S410-042