Implement accesses to one array element
trans_dest is mutually recursive with the translation of expression so
move it. It can also generate stmts since we need to evaluate the index
for an array access.

We only take into account accesses to one element of an array.

update lvalue_type_expression to also take CallExpr into account to
return the type of the element.

TN: S410-042