Implement array declaration
In Ada, when an array is declared, it is implicitely allocated. Thus, we
call malloc whenever an array is declared. For now we call malloc, but
in long term, it would be good to implement our own model of array
declaration instead of relying on malloc.

Also implement the constraints given by the type of an array.

TN: S410-042